requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==1.3.0
namex==0.0.8
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sympy==1.13.1
systemd-python==234
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs @ https://pypi.nvidia.com/tensorrt-libs/tensorrt_libs-8.6.1-py2.py3-none-manylinux_2_17_x86_64.whl#sha256=b8445cdba68d108345c95a65167c2bb6e03cb3e6cd6cb51e86f1028151d5a93e
termcolor==2.5.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.1.3
wrapt==1.14.1
zipp==1.0.0
zope.interface==5.4.0
